requirements.txt
Django==1.6.1
djangorestframework==2.3.12
wsgiref==0.1.2
